requirements.txt
allure-pytest==2.12.0
PyHamcrest==2.0.4
pytest==7.2.0
requests==2.28.1
selenium==4.7.2
webdriver-manager==3.8.5
pytest-html==3.2.0
pytest-rerunfailures==10.3